Rename and move workflow commands
We have other non-workflow commands that should not derive from CommandBase.
