+++ src/main/java/org/ilastik/ilastik4ij/workflow/ObjectClassificationCommand.java
@@ -1,4 +1,4 @@
-package org.ilastik.ilastik4ij.ui;
+package org.ilastik.ilastik4ij.workflow;
 
 import net.imagej.Dataset;
 import net.imglib2.type.NativeType;
@@ -10,7 +10,7 @@ import java.util.Collections;
 import java.util.Map;
 
 @Plugin(type = Command.class, headless = true, menuPath = "Plugins>ilastik>Run Object Classification Prediction")
-public final class ObjectClassificationCommand<T extends NativeType<T>> extends CommandBase<T> {
+public final class ObjectClassificationCommand<T extends NativeType<T>> extends WorkflowCommand<T> {
     @Parameter(label = "Pixel Probability or Segmentation image")
     public Dataset inputProbOrSegImage;
