+++ src/main/java/org/ilastik/ilastik4ij/workflow/WorkflowCommand.java
@@ -1,10 +1,11 @@
-package org.ilastik.ilastik4ij.ui;
+package org.ilastik.ilastik4ij.workflow;
 
 import net.imagej.Dataset;
 import net.imagej.ImgPlus;
 import net.imglib2.type.NativeType;
 import org.ilastik.ilastik4ij.hdf5.Hdf5DataSetReader;
 import org.ilastik.ilastik4ij.hdf5.Hdf5DataSetWriter;
+import org.ilastik.ilastik4ij.ui.IlastikOptions;
 import org.scijava.ItemIO;
 import org.scijava.app.StatusService;
 import org.scijava.command.ContextCommand;
@@ -38,7 +39,7 @@ import java.util.function.Consumer;
  * Subclasses should be named as {@code [MyWorkflowName]Command}
  * because workflow names are derived from their corresponding class names.
  */
-public abstract class CommandBase<T extends NativeType<T>> extends ContextCommand {
+public abstract class WorkflowCommand<T extends NativeType<T>> extends ContextCommand {
     public static final String ROLE_PROBABILITIES = "Probabilities";
     public static final String ROLE_SEGMENTATION = "Segmentation";
